+++ rangeforce.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2019, Matjaž Guštin <dev@matjaz.it> <https://matjaz.it>.
+# Copyright © 2019-2020, Matjaž Guštin <dev@matjaz.it> <https://matjaz.it>.
 # Released under the BSD 3-Clause License
 
 """Rangeforce, a developer-friendly check for "is this value
@@ -17,16 +17,18 @@ For example:
         value = int(input('How many hours per day do you sleep? '))
         value = rangeforce.limited(value, 0, 24, name='Hours of sleep')
         # Now value is valid. Otherwise an error message like this appears:
-        # rangeforce.RangeError: Hours of sleep must be in range [0, 24]. 25 found instead.
+        # rangeforce.RangeError: Hours of sleep must be in range [0, 24]. 25
+        found instead.
 
-        # Expecially useful for values that need to fit within an integer type:
+        # Especially useful for values that need to fit within an integer type:
         value = rangeforce.uint16(int(input('Type a 16-bit value: ')))
-        rangeforce.RangeError: Value must be in range [0, 65535]. 70000 found instead.
+        rangeforce.RangeError: Value must be in range [0, 65535]. 70000
+        found instead.
 """
 
 import math
 
-__VERSION__ = '1.0.0'
+__VERSION__ = '1.1.0'
 
 
 class RangeError(Exception):
@@ -62,11 +64,86 @@ def clip(value, min, max):
         return value
 
 
-def limited(value, min, max, name='Value', dtype=None):
+def exactly(value, expected, name='Value', dtype=None, ex=RangeError):
+    """Validates that value is exactly equal to another one.
+
+    If the value is valid, it returns the value itself: this is also the case
+    for comparing NaN to NaN, to avoid having a separate function for it.
+    In other words: exactly(NaN, NaN) is valid, does not raise exceptions.
+
+    If the value is not valid, it raises an exception of type ex with an
+    understandable error message that includes expected and failing value.
+
+    The name of the value can be altered for a customized error message.
+
+    The data type can be enforced if specified.
+
+    The exception class can be altered to a custom one.
+
+    Args:
+       value: the value to be validated to be within [min, max]
+       expected: only acceptable value. Not None. Can be NaN.
+       name: customizable name of the value that appears in the error message
+       dtype: optional data type the value has to be
+       ex: exception type to throw in case the value is out of range
+
+    Returns:
+       the given value if matching the expected value and, optionally, if of
+       the correct data type
+
+    Raises:
+       RangeError or type(ex): if the value is not matching the expected one.
+       TypeError: if the value is not of the acceptable data type, if
+                  specified.
+
+    Examples:
+           >>> exactly(0.5, 0.5)  # Valid value
+           0.5
+           >>> exactly(500, 30)  # Incorrect value
+           rangeforce.RangeError: Value must be exactly 30. 500
+           found instead.
+           >>> exactly(50, 7, name='Days in a week')
+           rangeforce.RangeError: Days in a week must be exactly 7. 50 found
+           instead.
+           >>> exactly(42.0, math.nan)
+           rangeforce.RangeError: Value must be exactly NaN. 42.0
+           found instead..
+           >>> exactly(math.nan, math.nan)  # Valid equality, also for NaN
+           nan
+           >>> exactly(7.0, 7, name='Days in a week', dtype=int)
+           TypeError: Days in a week must be of type int. float found
+           instead.
+           >>> exactly(7.0, 7, name='Days in a week', ex=ValueError)
+           ValueError: Days in a week must be exactly 7. 50 found instead.
+    """
+    _validate_type(name, value, dtype)
+    try:
+        if math.isnan(expected):
+            if math.isnan(value):
+                # NaN is EQUAL to NaN for this function.
+                return value
+            else:
+                raise ex(
+                    '{:} must be exactly NaN. '
+                    '{:} found instead.'.format(name, value)
+                )
+    except TypeError:
+        # Suppress math.isnan() applied to non-floats and just go on.
+        pass
+    if value != expected:
+        raise ex(
+            '{:} must be exactly {:}. '
+            '{:} found instead.'.format(name, expected, value)
+        )
+    else:
+        return value
+
+
+def limited(value, min, max, name='Value', dtype=None, ex=RangeError):
     """Validates that value is within the [min, max] interval.
 
     If the value is valid, it returns the value itself.
-    If the value is not valid, it raises a RangeError with an understandable
+    If the value is not valid, it raises an exeption with an understandable
     error message that includes expected range and failing value.
 
     Either min or max can be set to None for an unbound validity interval, i.e.
@@ -77,6 +154,8 @@ def limited(value, min, max, name='Value', dtype=None):
 
     The data type can be enforced if specified.
 
+    The exception class can be altered to a custom one.
+
     Args:
         value: the value to be validated to be within [min, max]
         min: smallest acceptable value. Can be None if max is not None.
@@ -85,13 +164,15 @@ def limited(value, min, max, name='Value', dtype=None):
              Can be +inf, -inf. Cannot be NaN. Must be >= min.
         name: customizable name of the value that appears in the error message
         dtype: optional data type the value has to be
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value if within [min, max] and, optionally, of the correct
         data type
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not of the acceptable data type, if
                    specified.
         ValueError: if the min, max extremes are not valid (e.g. both None,
@@ -104,31 +185,34 @@ def limited(value, min, max, name='Value', dtype=None):
             rangeforce.RangeError: Value must be in range [0.1, -42]. 500
             found instead.
             >>> limited(50, 0, 24, name='Hours in a day')
-            rangeforce.RangeError: Hours in a day must be in range [0,
-            24]. 50 found instead.
+            rangeforce.RangeError: Hours in a day must be in range [0, 24]. 50
+            found instead.
             >>> limited(-1, 0, None, name='Earth satellites')
-            rangeforce.RangeError: Earth satellites must be in range [0,
-            +inf[. -1 found instead.
+            rangeforce.RangeError: Earth satellites must be in range [0, +inf[.
+            -1 found instead.
             >>> limited(1.1, 0, None, name='Earth satellites', dtype=int)
             TypeError: Earth satellites must be of type int. float found
+            instead.
+            >>> limited(1.1, 0, None, name='Earth satellites', ex=ValueError)
+            ValueError: Earth satellites must be in range [0, +inf[. -1 found
             instead.
     """
     _validate_interval(min, max)
     _validate_type(name, value, dtype)
     if min is None and max is not None and (value > max or math.isnan(value)):
-        raise RangeError(
+        raise ex(
             '{:} must be in range ]-inf, {:}]. '
             '{:} found instead.'.format(name, max, value)
         )
     elif max is None and min is not None and (
             value < min or math.isnan(value)):
-        raise RangeError(
+        raise ex(
             '{:} must be in range [{:}, +inf[. '
             '{:} found instead.'.format(name, min, value)
         )
     elif min is not None and max is not None and (
             value < min or value > max or math.isnan(value)):
-        raise RangeError(
+        raise ex(
             '{:} must be in range [{:}, {:}]. '
             '{:} found instead.'.format(name, min, max, value)
         )
@@ -158,7 +242,7 @@ def _validate_type(name, value, dtype):
         )
 
 
-def negative_int(value, name='Value'):
+def negative_int(value, name='Value', ex=RangeError):
     """Validates that value is negative (< 0) and of type int.
 
     If the value is valid, it returns the value itself.
@@ -168,18 +252,20 @@ def negative_int(value, name='Value'):
     Args:
         value: the value to be validated to be within ]-inf, 0[
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value if < 0
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, None, -1, name, dtype=int)
+    return limited(value, None, -1, name, dtype=int, ex=ex)
 
 
-def nonpositive_int(value, name='Value'):
+def nonpositive_int(value, name='Value', ex=RangeError):
     """Validates that value is non-positive (<= 0) and of type int.
 
     If the value is valid, it returns the value itself.
@@ -189,18 +275,20 @@ def nonpositive_int(value, name='Value'):
     Args:
         value: the value to be validated to be within ]-inf, 0]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value if <= 0
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, None, 0, name, dtype=int)
+    return limited(value, None, 0, name, dtype=int, ex=ex)
 
 
-def positive_int(value, name='Value'):
+def positive_int(value, name='Value', ex=RangeError):
     """Validates that value is positive (> 0) and of type int.
 
     If the value is valid, it returns the value itself.
@@ -210,18 +298,20 @@ def positive_int(value, name='Value'):
     Args:
         value: the value to be validated to be within ]0, +inf[
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value if > 0
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, 1, None, name, dtype=int)
+    return limited(value, 1, None, name, dtype=int, ex=ex)
 
 
-def nonnegative_int(value, name='Value'):
+def nonnegative_int(value, name='Value', ex=RangeError):
     """Validates that value is non-negative (>= 0) and of type int.
 
     If the value is valid, it returns the value itself.
@@ -231,18 +321,20 @@ def nonnegative_int(value, name='Value'):
     Args:
         value: the value to be validated to be within [0, +inf[
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value if >= 0
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, 0, None, name, dtype=int)
+    return limited(value, 0, None, name, dtype=int, ex=ex)
 
 
-def uint8(value, name='Value'):
+def uint8(value, name='Value', ex=RangeError):
     """Validates that value fits in an 8-bit unsigned integer.
 
     If the value is valid, it returns the value itself.
@@ -252,18 +344,20 @@ def uint8(value, name='Value'):
     Args:
         value: the value to be validated to be within [0, 255]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in 8 bits (unsigned)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, 0, 0xFF, name, dtype=int)
+    return limited(value, 0, 0xFF, name, dtype=int, ex=ex)
 
 
-def uint16(value, name='Value'):
+def uint16(value, name='Value', ex=RangeError):
     """Validates that value fits in a 16-bit unsigned integer.
 
     If the value is valid, it returns the value itself.
@@ -273,18 +367,20 @@ def uint16(value, name='Value'):
     Args:
         value: the value to be validated to be within [0, 65535]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in 16 bits (unsigned)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, 0, 0xFFFF, name, dtype=int)
+    return limited(value, 0, 0xFFFF, name, dtype=int, ex=ex)
 
 
-def uint32(value, name='Value'):
+def uint32(value, name='Value', ex=RangeError):
     """Validates that value fits in a 32-bit unsigned integer.
 
     If the value is valid, it returns the value itself.
@@ -294,18 +390,20 @@ def uint32(value, name='Value'):
     Args:
         value: the value to be validated to be within [0, 4294967295]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in 32 bits (unsigned)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, 0, 0xFFFFFFFF, name, dtype=int)
+    return limited(value, 0, 0xFFFFFFFF, name, dtype=int, ex=ex)
 
 
-def uint64(value, name='Value'):
+def uint64(value, name='Value', ex=RangeError):
     """Validates that value fits in a 64-bit unsigned integer.
 
     If the value is valid, it returns the value itself.
@@ -315,18 +413,20 @@ def uint64(value, name='Value'):
     Args:
         value: the value to be validated to be within [0, 18446744073709551615]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in 64 bits (unsigned)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, 0, 0xFFFFFFFFFFFFFFFF, name, dtype=int)
+    return limited(value, 0, 0xFFFFFFFFFFFFFFFF, name, dtype=int, ex=ex)
 
 
-def uint_bits(value, bits, name='Value'):
+def uint_bits(value, bits, name='Value', ex=RangeError):
     """Validates that value fits in an unsigned integer of specified bitlength.
 
     If the value is valid, it returns the value itself.
@@ -335,20 +435,24 @@ def uint_bits(value, bits, name='Value'):
 
     Args:
         value: the value to be validated to be within [0, 2**bits-1]
+        bits: integer, positive number of bits determining the size and thus
+              the range of the value
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in the given amount of  bits
         (unsigned)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, 0, (1 << bits) - 1, name, dtype=int)
+    return limited(value, 0, (1 << bits) - 1, name, dtype=int, ex=ex)
 
 
-def int8(value, name='Value'):
+def int8(value, name='Value', ex=RangeError):
     """Validates that value fits in an 8-bit signed integer.
 
     If the value is valid, it returns the value itself.
@@ -358,18 +462,20 @@ def int8(value, name='Value'):
     Args:
         value: the value to be validated to be within [-128, 127]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in 8 bits (signed)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, -0x80, 0x7F, name, dtype=int)
+    return limited(value, -0x80, 0x7F, name, dtype=int, ex=ex)
 
 
-def int16(value, name='Value'):
+def int16(value, name='Value', ex=RangeError):
     """Validates that value fits in a 16-bit signed integer.
 
     If the value is valid, it returns the value itself.
@@ -379,18 +485,20 @@ def int16(value, name='Value'):
     Args:
         value: the value to be validated to be within [-32768, 32767]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in 16 bits (signed)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, -0x8000, 0x7FFF, name, dtype=int)
+    return limited(value, -0x8000, 0x7FFF, name, dtype=int, ex=ex)
 
 
-def int32(value, name='Value'):
+def int32(value, name='Value', ex=RangeError):
     """Validates that value fits in a 32-bit signed integer.
 
     If the value is valid, it returns the value itself.
@@ -400,18 +508,20 @@ def int32(value, name='Value'):
     Args:
         value: the value to be validated to be within [-2147483648, 2147483647]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in 32 bits (signed)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
-    return limited(value, -0x80000000, 0x7FFFFFFF, name, dtype=int)
+    return limited(value, -0x80000000, 0x7FFFFFFF, name, dtype=int, ex=ex)
 
 
-def int64(value, name='Value'):
+def int64(value, name='Value', ex=RangeError):
     """Validates that value fits in a 16-bit signed integer.
 
     If the value is valid, it returns the value itself.
@@ -422,23 +532,25 @@ def int64(value, name='Value'):
         value: the value to be validated to be within [-9223372036854775808,
                9223372036854775807]
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given value can be expressed in 64 bits (signed)
 
     Raises:
-        RangeError: if the value is not within the acceptable range.
+        RangeError or type(ex): if the value is not within the acceptable
+                                range.
         TypeError: if the value is not an integer.
     """
     return limited(value, -0x8000000000000000, 0x7FFFFFFFFFFFFFFF, name,
-                   dtype=int)
+                   dtype=int, ex=ex)
 
 
-def limited_len(sized, min, max, name='value'):
+def limited_len(sized, min, max, name='value', ex=RangeError):
     """Validates that value has a length within the [min, max] interval.
 
     If the sized value is valid, it returns the value itself.
-    If the sized value is not valid, it raises a RangeError with an
+    If the sized value is not valid, it raises an exception with an
     understandable error message that includes expected length range and
     failing sized value.
 
@@ -448,6 +560,8 @@ def limited_len(sized, min, max, name='value'):
 
     The name of the sized value can be altered for a customized error message.
 
+    The exception class can be altered to a custom one.
+
     Args:
         sized: the value whose length is to be validated to be within
                [min, max]
@@ -456,13 +570,14 @@ def limited_len(sized, min, max, name='value'):
         max: greatest acceptable length. Can be None if min is not None.
              Can be +inf, -inf. Cannot be NaN. Must be >= min and >= 0.
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given sized value if has length within [min, max]
 
     Raises:
-        RangeError: if the value does not have a length within the acceptable
-                    range.
+        RangeError or type(ex): if the value does not have a length within the
+                                acceptable range.
         ValueError: if the min, max extremes are not valid (e.g. negative,
                     both None, min greater than max, NaN etc.)
 
@@ -478,10 +593,12 @@ def limited_len(sized, min, max, name='value'):
             >>> limited_len([1, 2, 3], 10, None)
             rangeforce.RangeError: Length of value must be in range [10,
             +inf[. 3 found instead.
+            >>> limited_len([1, 2, 3], 10, 20, ex=ValueError)
+            ValueError: Length of value must be in range [10, 20]. 3 found
+            instead.
     """
-    length = len(sized)
     _validate_non_negative_interval_extremes(min, max)
-    limited(length, min, max, name='Length of ' + name)
+    limited(len(sized), min, max, name='Length of ' + name, ex=ex)
     return sized
 
 
@@ -498,29 +615,33 @@ def _validate_non_negative_interval_extremes(min, max):
         )
 
 
-def exact_len(sized, expected, name='value'):
+def exact_len(sized, expected, name='value', ex=RangeError):
     """Validates that value has an exact length.
 
     If the sized value is valid, it returns the value itself.
-    If the sized value is not valid, it raises a RangeError with an
-    understandable error message that includes expected length and
-    failing sized value.
+    If the sized value is not valid, it raises an exception with an
+    understandable error message that includes expected length and failing
+    sized value.
 
     The name of the sized value can be altered for a customized error message.
+
+    The exception class can be altered to a custom one.
 
     Args:
         sized: the value whose length is to be validated to be exactly as
                expected.
         expected: only acceptable length. Must be an integer >= 0.
         name: customizable name of the value that appears in the error message
+        ex: exception type to throw in case the value is out of range
 
     Returns:
         the given sized value if has length matching the expected
 
     Raises:
-        RangeError: if the value does not have a length matching the expected.
-        ValueError: if the expected length is not valid (e.g. not integer,
-                    negative, None)
+        RangeError or type(ex): if the value does not have a length matching
+                                the expected.
+        TypeError: if the expected length is not an integer
+        ValueError: if the expected length is negative
 
     Examples:
             >>> exact_len([1, 2, 3], 3)  # Valid value
@@ -531,11 +652,13 @@ def exact_len(sized, expected, name='value'):
             >>> exact_len([1], 2, name='pairs')
             rangeforce.RangeError: Length of pairs must be exactly 2. 1
             found instead.
+            >>> exact_len([1], 2, name='pairs', ex=ValueError)
+            ValueError: Length of pairs must be exactly 2. 1 found instead.
     """
     length = len(sized)
     _validate_expected_length(expected)
     if length != expected:
-        raise RangeError(
+        raise ex(
             'Length of {:} must be exactly {:}. '
             '{:} found instead.'.format(name, expected, length)
         )