+++ CHANGELOG.md
@@ -9,6 +9,21 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[1.1.0] - 2020-07-08
+----------------------------------------
+
+### Added
+
+- `exactly(value, expected)` checking if a value is equal to the expected one.
+  This behaves similarly to `assertEqual()`. It also evaluates NaN == NaN
+  as True for practical reasons (details in function docs). 
+- Parameter `ex` to all functions to customise the exception type.
+  Defaults to `RangeError`, thus keeping backwards compatibility, but the user
+  can choose to use `ValueError`, `OverflowError`, `FileNotFoundError`
+  or any other custom exception class.
+
+
+
 [1.0.0] - 2019-04-20
 ----------------------------------------